% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_sf.R
\name{arc_sf}
\alias{arc_sf}
\title{Title}
\usage{
arc_sf(.dsn, .query = NULL)
}
\arguments{
\item{.dsn}{The path to the data resource (Living Atlas)}

\item{.query}{Tne quoted query to subset data}
}
\value{
a spatial data frame (sf)
}
\description{
Title
}
