% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lns.R
\name{get_lns}
\alias{get_lns}
\title{Fetch spatial lines from the UCRB rivers dataset}
\usage{
get_lns(.where = NULL)
}
\arguments{
\item{.where}{A SQL style WHERE statement}
}
\value{
A simple features dataframe
}
\description{
Fetch spatial lines from the UCRB rivers dataset
}
\examples{
get_lns(.where = "cd_rvr IN ('GR', 'PR', 'WH')") |>
(\(x) {mapview::mapview(x)}) ()
}
