% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_pts.R
\name{map_pts}
\alias{map_pts}
\title{Map Points}
\usage{
map_pts(.x, .y, .crs = 4326)
}
\arguments{
\item{.x}{The x coordinate(s). Use the combine operator (c) when supplying multiple values}

\item{.y}{The y coordinate(s). Use the combine operator (c) when supplying multiple values}

\item{.crs}{The EPSG code for the coordinates. Defaults to 4326 (WGS84)}
}
\value{
Mapview interactive web map
}
\description{
Map Points
}
\examples{
x <- c(-109, -109.5, -110)
y <- c(39, 39, 39)
map_pts(.x = x, .y = y)
}
