% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_rmi.R
\name{map_rmi}
\alias{map_rmi}
\title{Specify locations by river code and mile and visualize them on map}
\usage{
map_rmi(.cd_rvr, .rmi_nhd)
}
\arguments{
\item{.cd_rvr}{A UCRB Program river code}

\item{.rmi_nhd}{A river mile based on NHD data}
}
\value{
An interactive map
}
\description{
Specify locations by river code and mile and visualize them on map
}
\examples{
map_rmi(.cd_rvr = "GR", .rmi_nhd = c(42, 52, 120))
}
