% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_rvr.R
\name{map_rvr}
\alias{map_rvr}
\title{Map river(s)}
\usage{
map_rvr(.where, .points = TRUE)
}
\arguments{
\item{.where}{A SQL style WHERE statement}

\item{.points}{Do you want 10th river mile points mapped (TRUE) or centerlines (FALSE)}
}
\value{
A mapview interactive map
}
\description{
Map river(s)
}
\examples{
map_rvr(.where = "cd_rvr IN ('GU', 'CO')",  .points = FALSE)
map_rvr(.where = NULL, .points = FALSE)
}
